Conflicts:
	Codename Frozen Omega/src/CodenameFrozenOmega/ObjectsControl.java
	Codename Frozen Omega/src/CodenameFrozenOmega/Ticktick.java

+++ Codename Frozen Omega/src/CodenameFrozenOmega/Herbivore.java
@@ -2,7 +2,7 @@ package CodenameFrozenOmega;
 
 public class Herbivore extends Animal {
 
-	protected Herbivore(int preferredTemperature, int id, int location, int lifespan, double age, String species,
+	protected Herbivore(int preferredTemperature, int id, int location, double lifespan, double age, String species,
 			int foodChainPosition, int habitability, int birthTick, boolean healthy) {
 		super(preferredTemperature, id, location, lifespan, habitability, age, species, foodChainPosition, birthTick, healthy);
 	}
